Update function works (and works better?)
Almost everything is done in origr.makeMagic(), including
checking if target got removed from the DOM (it's being stored in
origr.currentTarget, so it's not completely gone and a special
check has to be made)
It's all getting called from origr.update(), though, which
makes the event listener less clear.
Not sure what to do about that.